Exit on unexpected params; better testing when launching; use bash to create pid file.

+++ xwords4/relay/xwrelay.sh
@@ -3,32 +3,36 @@
 XWRELAY="./xwrelay"
 PIDFILE=./xwrelay.pid
 
+CMD=$1
+shift
 
-case $1 in
+case $CMD in
     
     stop)
-        if [ -f $PIDFILE ]; then
+        if [ -f $PIDFILE ] && [ -f /proc/$(cat $PIDFILE)/exe ]; then
             sync
             echo "killing pid=$(cat $PIDFILE)"
             kill $(cat $PIDFILE)
-            rm $PIDFILE
         else
             echo "not running"
         fi
+        rm $PIDFILE
         ;;
 
     restart)
         $0 stop
         sleep 1
-        $0 start
+        $0 start $@
         ;;
 
     start|*)
-        if [ -f $PIDFILE ]; then
+        if [ -f $PIDFILE ] && [ -f /proc/$(cat $PIDFILE)/exe ]; then
             echo "already running: pid=$(cat $PIDFILE)"
         else
             echo "starting..."
-            $XWRELAY
+            $XWRELAY $@ &
+            NEWPID=$!
+            echo $NEWPID > $PIDFILE
             sleep 1
             echo "running with pid=$(cat $PIDFILE)"
         fi
